[ETL/Core] Fixed warnings (#2013)
- static_cast<void*> used to disable warning "clearing an object of non-trivial type...". This can be removed after synfig::Color will be trivial;
- Removed unused variables;
- Fixed warning: comparison between signed and unsigned integer expressions.

+++ synfig-studio/src/synfigapp/vectorizer/centerlinepolygonizer.cpp
@@ -767,7 +767,7 @@ inline double penalty(RawBorder &path, int a, int b)
   synfig::PointInt p;
 
   int n = b - a + 1;
-  p = path[b == path.size() ? 0 : b].pos() - path[a].pos();
+  p = path[b == static_cast<std::int32_t>(path.size()) ? 0 : b].pos() - path[a].pos();
 
   synfig::Point v(-p[1],p[0]);//convert(rotate90(p))
   synfig::Point sum   = path.sums()[b] - path.sums()[a];
